Update Bioinformatics coursework
Inputs and outputs

The user has to run a single script from the command line, where he needs to enter a taxon/species name of interest. After the script has been executed, three files are generated and appear in the working directory. The first file is a FASTA file named efetch_<Taxon name>.fasta that contains single–line descriptions followed by the amino acid (aa) sequences of proteins. The second file - the EMBOSS Pepstats output with different calculated statistics for each protein, and is named as pepstats_output_<Taxon name>.pepstats. The final file is a tab-delimited text file named AAcid_table_of_<Taxon name >.txt that provides a simple table with three columns. The first column - the single-letter abbreviations for the common and rare aas, the second column - the count of those aas, and the third column - the percentage (molar %) of aa occurrence. The user will be able to easily import the file with the table into Microsoft Excel and access additional information if required. 

Summary of the pipeline 

•	The first step involves a read command, followed by an if-else statement. Firstly, read reads the input - a name of taxon/species provided by the user. If nothing is entered, then the user is asked again to provide the taxon/species name. The user has an additional option to terminate the shell script. 
•	Next step involves wget, grep, cut and the if-else statement. If the correct taxon/species name was provided, then it will be used as a major search term together with Boolean OR/NOT operators (used to filter the search) that will form a search query. It will be used for research in GenBank to find all proteins. The output of esearch will be stored in a text file protein_<Taxon name>.txt and will contain a list of unique identifiers (UIDs), including a temporary Web environment string (WebEnv), which contains the results of esearch. grep commands then use  <”webenv”>  and   <MegaStore">  to extract the full WebEnv string that is piped to wc – l, which counts the number of lines identified. Following, if else statement checks if a text file is correct and contains WebEnv; otherwise, the script is terminated. As before, grep extracts the full WebEnv by <”webenv”>. cut in combination with echo, remove unnecessary characters. The processed string is then stored in a variable that is used in another wget command that employs the EFetch utility to obtain the FASTA files from GenBank. 
•	grep and if else statements test if the FASTA file - efetch_< Taxon name>.fasta is correct. grep searches for <ERROR> and if the file has this pattern, wc -l counts the number of lines with <ERROR> and stores in a fasta_line variable. It is then utilised by the while loop to verify if the FASTA file is correct. If it is empty, the script is terminated. 
•	awk counts and sums the number of lines containing ‘ > ‘. If else statement then informs the user if all or the first 10,000 protein sequences are downloaded and stored in efetch_< Taxon name>.fasta
•	The FASTA file is applied to the EMBOSS Pepstats tool that reads protein sequences and calculates statistics, which is stored in pepstats_output_<Taxon name>.pepstats. Then a series of awk commands searches for the pattern similar to ‘ A = Ala ‘ for each aa and sums each aa across all proteins. The results are stored in separate variables and saved in a text file named file_count.txt. The same process is used to sum the percentage occurrence of each aa that is saved in file_percentage.txt. dc -f uses both files to estimate the average percentages of occurrence of each aa across all proteins that are saved in average_percentage.txt. A for loop then generates a text file acid_names.txt with the single-letter abbreviations of all aas. 
•	paste and awk merge acid_names.txt, file_count.txt and file_percentage.txt, and create a single tab-delaminated file called merged_file.txt. Then, sort -k2 and sed -i are applied to sort merged_file.txt based on ‘Counts’ in descending order and to insert a top row with column names that is saved as a text file called AAcid_table_of_<Taxon name>.txt
•	awk provides the sum of sequences, while the for loop utilises file_count.txt to calculate the total number of aas used. echo commands provide these results to the user. rm commands are used to delete intermediate files. 

Instructions

The script has to be run in a Linux terminal from the command line. Initially, it will ask the user to give the taxon/species name. There is no character limit; therefore, the provided input can be a generic name, Latin name or taxonomy ID. However, if the taxon/species name consists of two words, they should be separated by an underscore. The output files can be generated for one species at a time. The user also has an option to terminate the script. If the correct name is provided, three different files will be generated: the FASTA file, the pepstats output and the text file with the summary table of aa usage of the taxon/species. Note, if the script is terminated due to an incorrect FASTA file, the taxonomy ID should be given as input. 

Examples of use

Two selected species were a home cat (Felis catus) and an apple (Malus domestica). The provided inputs were cat and txid3750. These species were selected as they are evolutionary distant and are well-known species. The AAcid_table_of_<Taxon name>.txt files were created for both species (Table 1). The top five most frequent aas were leucine, serine, alanine and glutamic acid, and proline in Felis catus or glycine in Malus domestica. Thus, most frequent aas are similar even across very distant species, but in terms of percentage occurrence, these aas are more frequent in Malus domestica than compared to Felis catus. For Felis catus, 10,000 sequences and 7,330,924 aas were used, while for Malus domestica, in total 10,000 sequences and 4,799,514 aas were used. 

Table 1. The summary table of amino acid usage of the two species

cat (Felis catus)		txid3750 (Malus domestica)
Amino acid	Count	Percentage		Amino acid	Count	Percentage
L	721743	0.136		L	459965	0.204
S	610231	0.132		S	440101	0.205
A	554877	0.139		A	326522	0.211
E	540936	0.127		E	318125	0.203
P	512900	0.137		G	314926	0.211

Evaluation and future directions

The provided pipeline is fully automated and, therefore, it is user-friendly. Also, it is relatively robust to errors and provides the output files, thus fulfilling its purpose. However, it has many limitations. Firstly, if an incorrect taxon name is entered (characters, numbers), a wrong FASTA file is generated, and due to the pipeline is terminated. However, when the symbols (‘+’, ‘-‘, ’.’, ’!’) are entered, the pipeline generates full output files.
Similarly, it may not produce the output files for rarer, less studied or extinct species, e.g., the giant ground sloth. Secondly, efetch has a limit, and therefore, this pipeline downloads a maximum of 10,000 protein sequences. Some sequences may repeat, and this pipeline does not exclude duplicated sequences. This could be achieved with a while/for loop and applying grep, awk and some other commands, possibly an if statement. The first description line with sequence ID (FASTA file) could be extracted with grep/awk and used as a reference. Then, all sequences that are not identical to the reference could be stored in a text file. This could be repeated with the second and following description lines, and the text file, storing selected sequences, could be overwritten every time, in this way, leaving only those sequence IDs that are unique. Finally, the script has many lines of code that are repeating, e.g. lines with awk. They could be replaced or incorporated into a for loop, as multiple repetitions can be prone to errors.